fix: provide GatewayBuilder with an actual partial_cache_registry (#1809)
I was trying to integration test partial caching, but the partial cache
registry inside the tests always had partial caching disabled, no matter
what I put into my schema. This turned out to be because it was always
hardcoded to empty.

This PR fixes that - instead of constructing a GatewayBuilder directly
we construct it from the EngineBuilder, which allows us to pass through
the correct partial caching registry.