chore(engine-v2): Engine generic over the runtime & hooks trait update
Was a bit tedious to set up and encountered a bug with Rust compiler but
now engine is generic over its runtime. This allow us to have
considerably more flexibility in what we can do in our interfaces.

I've also updated the hook trait to be closer to the next RFC, but only
added the `pre-execution` hooks for now. I've simplified the error
handling also, by having a single `runtime::error::GraphqlError` instead
of `HookError` and `UserError`. This `GraphqlError` only has message and
extension, the rest will be added by the engine. The distinction between
a HookError & UserError isn't necessary for the engine, in all cases
it's treated as an error that must be shown.

To avoid re-compiling the engine for various runtimes in tests, I've
added a `TestRuntime` struct and added a `DynHooks` trait with default
implementation which allows providing dynamically dispatched hooks.