feat: hook up partial caching defer work (#1810)
I've spent the past couple of weeks working on defer partial caching
without actually hooking it up to anything. I'd written most of the code
I'd needed, but until it was half way functional I didn't want to call
it from the gateway.

This PR (mostly) changes that:

- Added a new `execute_stream_v2` entrypoint to the Gateway. This is
very similar to the old `execute_stream` function, but it returns a
`Stream<StreamingPayload>` rather than an http response. This allows
partial caching to work against this stream instead of having to deal
with the streaming transports that were contained within the http
response.
- This function calls into a new `StreamingExecutionPhase` struct, which
is similar to the `ExecutionPhase` used by non-streaming transports.
- The `gateway_core` interface also now has an `execute_stream_v2`
function.

Once all of this is implemented and hooked up I'll retire the
`execute_stream` functions in favor of v2, but that won't be until next
week at the earliest.

I've tested this with some integration tests that'll be in my next PR.
There are issues with it, but some of the basics work at least. It's not
yet usable outside of integration tests - I need to implement transports
in the CLI & cloud gateway first.

Fixes GB-6746