test: basic integration test of partial caching (#1811)
This adds a very simple integration test of partial caching w/ defer.
Had to add support for streaming requests to the gateway testing infra
for this, but nothing too complicated.

This is just a really simple test - I don't want to do anything more
in-depth, because I know there's a lot of broken things, and I want to
get the transport code written for the demo first. I've made a linear
issue to ensure I revisit once more things are working.