feat: impl streaming transports in cli gateway (#1814)
Prior to the partial caching defer work encoding streaming transports
was done in the executor, with the executor just returning an HTTP
response back to the gateway. Since partial caching needs access to the
responses, this is no longer the case. Instead we need to encode the
streaming transports at the http level. This is arguably a more sensible
approach anyway, that'll probably clean up some mess.

This PR takes care of doing that for the CLI, doing that in production
will come later.

Fixes GB-6985