feat(engine-v2): Add back operation caching
Adding a new trait `HotCache` to abstract away the in memory caching
with a simpler interface than the current `Cache`. Using it to store
Operation data and trusted documents & APQ.

The trait is designed to ensure we can use it in Cloudflare and with
in-memory cache (mini-moka here).

Locally I have the following results with the simple benchmarks I made:

| HotCache   |     basic | introspection |
| :--------- | --------: | ------------: |
| ()         |     38 us |        336 us |
| postcard*  |   15.8 us |        210 us |
| in-memory  |   15.4 us |        185 us |

*in-memory + postcard serialization

So should be also be of help in Cloduflare.