fix: support defers without labels in partial caching (#1825)
While writing the partial caching defer support, I made the assumpption
that defers always have labels. That was incorrect, they can be
unlabelled.

This updates the code to account for that: instead of using the defer
label to refer to defers we parse them out as part of the planning
phase, assign them ids and then use those ids to refer to them.

One place this got complicated is when we receive deferred payloads for
a defer without a label: we need to figure out which defer it's
referring to based on the path. This is relatively easy if there's only
one defer inside an object, but gets complicated if there's more than
one. Managed it in the end though.

Fixes GB-6981