chore: update wasmtime crates to 22.0.0 (#1840)
https://github.com/bytecodealliance/wasmtime/releases/tag/v22.0.0

This release has a regression preventing linking if compiling against
musl. I bisected and found the issue in wasmtime that caused this. What
we do here is we remove the three lines in wasmtime build.rs which link
against libm, and we have a working build.

Context: https://github.com/bytecodealliance/wasmtime/issues/8898
Changes against wasmtime 22:
https://github.com/grafbase/wasmtime/commit/e55186fe62b56ceb5a168a79ecd773dea6bdcca6

Closes: GB-7038