feat: support type conditions in partial caching (#1862)
Prior to this commit we'd just panic if we found a polymorphic shape
when we were handling a partial caching response. This commit fixes
that: we now use walk the possible type conditions looking for the most
specific match, and then use the corresponding concrete type.

I've also added a bunch of integration tests, which are the bulk of this
PR.

One thing I haven't done yet is update the cache writing code to write
type names into the cache entry. I'll do that in my next PR, as this one
is already quite large.

Fixes GB-6949