fix: write typename to partial caches (#1864)
When partial caching encounters a type condition, it will add a
`__typename` field into the query if one is not already present. These
`__typenames` are not written into the `OutputStore` as actual
`__typename` fields, so don't currently get written to the cache. But
this means when we read from the cache we don't know which type that
object is meant to be.

This commit fixes that by storing the typename in a new `typename` field
on the `Object`. We intern these to save on space, and only write them
if the query doesn't have a user provided `__typename` on it.

Now that this is in place I've updated the associated tests to hit the
cache and make sure that works.

Fixes GB-7037