fix: use `CompactValue` directly inside `StreamingPayload` (#1876)
I've been updating the `api` repo to use `execute_stream_v2` instead of
`execute_stream`. The main difference between these two is that `v2`
returns a `Stream<StreamingPayload>`, where as the original returned an
`http::Response.

In the API these functions need to make an HTTP call to another worker,
so I needed to serialise & deserialise the `StreamingPayload` on either
side of this HTTP call. This turned out to be quite hard, because
`StreamingPayload` didn't implement `Deserialize`. Implementing
`Deserialize` was also tricky, because it contained an
`engine::Response`, but didn't make use of the `engine::Response` impl
of `Serialize` but instead serialized to response format JSON. There's
no existing deserialize impl from response format JSON into an
engine::Response, and we also lose a bunch of info in the process so
it'd be hard to even write one. It's a big mess.

I tried 2 or 3 things to fix this, but the easiest one turned out to be
just removing `engine::Response` from `StremaingPayload` and having a
`CompactValue` instead. Which is what this commit does.

This might not be perfect - we normally ferry some things between the
gateway & the executor inside `engine::Response`, which will be
completely missing here. But I'll deal with the individual issues as I
discover them, want to get this out first and this was the easiest way
without refactoring all the things.