chore(engine-v2): authorized on types equivalent to applying on fields
Initially the idea was to deny access to an object whenever it was
present in the response. But that's hard to do with interfaces. And
makes `@authorized` on interfaces even harder and weird. So instead
I'm using a simpler logic that applying `@authorized` on a type is
equivalent to applying it on all of its fields. This does change its
meaning and I'll very likely change the hook definitions later, once
I've implemented all the logic.