chore(engine-v2): Create ResponseModifier
ResponseModifier are built in the same way QueryModifiers are during the
operation preparation. As their name suggests, they will modify the
response instead of modifying the query.

I've implemented most of the code for the response modifier in this PR:

- struct themselves
- being able to forward a view of the response objects to the
  hooks
- As a response modifier may be used on various selection sets but
  should be called at most once per subgraph response I
  created a `FilteredResponseObjectSet` which can iterate over multiple
  `ResponseObjectSet`
- I've renamed `ResponsePart` to be `SubgraphResponse` for hopefully a
  bit more clarity and tweaked a bit how executor receive it. Now they
  get a mutable reference to it allowing us to avoid a bunch of copies
  or the root response objects (where we need add the new fields) which
  I need in case of executor errors (to know where to add errors).

I've stopped there because defining the "ReadSelectionSet", what and how
we read from the response is a bit broken outside of plan root fields.
I'm not entirely sure what I'll do here, the best solution would be to
add the necessary metadata in the shapes.