chore(engine-v2): executor simplification
I realized what I was doing all along with executor is just a
hand-written future. So removing it entirely. Now there are only two
steps, instead of three:
- prepare
- execute

The reason for the intermediate step was that the response object view
which provides any required fields locks the response, and thus must be
used before awaiting anything. But that can be done just as well by
returning a future.