feat: entity caching for top level fields
This adds entity caching for requests to top level fields.  The name entity
caching isn't _really_ accurate for these, since they're not neccesarily
entities, but whatever.

The implementation is very barebones at the moment: we just hash the query, and
look that up in the cache.  We don't yet bother with auth scopes, cache
invalidation, or even much of the configuration concerns for this feature
(beyond what I needed for tests).