fix: expose otlp config structs regardless of otlp feature (#1932)
There's been problems pulling this repo into our `api` repo in the past
couple of days. This seems to be because the clickhouse tests require in
a bunch of structs that are hidden behind the `oltp` feature flag in the
`grafbase_telemetry` crate, and as soon as you enable that flag you pull
in a bunch of additional crates, that don't play well with the set of
dependencies we have in `api`, and unpicking that is nightmarish.

Prior to
https://github.com/grafbase/grafbase/commit/0f4cdd3bc2cfe92feeb77cb81bf548cddb35d87d
these structs were all exposed even when the `oltp` feature was turned
off, which allowed the clickhouse tests to compile without introducing
this dependency hell situation.

This commit removes the oltp feature flag from these structs so that the
clickhouse tests compile, while still keeping the oltp functionality
behind the feature flag. I'm not entirely sure if the clickhouse tests
_work_ with this setup, but they do seem to compile.