fix: dont set timeout to u32::MAX for streaming requests
The recent addition of timeouts set a timeout of u32::MAX for streaming
requests.  I assume this works fine on native, but on [wasm it gets cast to an
i32][1], so it ends up negative and we instantly timeout.  Better to just not
timeout at all when we get a streaming request.

[1]: https://github.com/rustwasm/gloo/blob/master/crates/timers/src/callback.rs#L66