fix: dont set timeout to u32::MAX for streaming requests (#1936)
The recent addition of timeouts set a timeout of u32::MAX for streaming
requests. I assume this works fine on native, but on [wasm it gets cast
to an i32][1], so it ends up negative and we instantly timeout. Better
to just not timeout at all when we get a streaming request.

This is causing e2e failures in the api repo - _hopefully_ this is the
last thing that needs fixed before we can finish pulling in the latest
grafbase over there.

[1]: https://github.com/rustwasm/gloo/blob/master/crates/timers/src/callback.rs#L66