fix(engine-v2): Panic could happen with both pre & post execution auth
The post-execution mechanism, `ResponseModifierExecutor`, was expecting
its input, a `ResponseObjectSet`, to exist once its parent plans were
executed. However with pre-execution auth we modify the subgraph query
to not retrieve any unauthorized fields. This could lead to the input
never being created. And as I'm writing this I realize it could have
happened in different circumstances, so well. Added a tests for it.

I'm also changing a bit how I'm testing type system directives. Before I
was adding them manually in the "secure" graphql-mock, but it's already
a mess today. So instead I created a simple TeaShop schema and I'm
overriding the SDL to add directives for a given test. Making it a lot
easier to follow and more re-usable. Will try to move the other hook
stuff to this model over time.