chore(integration-tests): Simplifying subgraph registration & received requests
Subgraph had to be created separately and kept alive during the tests
with a dummy subgraph name. Was pretty much boilerplate as it ended up
being just more or less well copy pasted everywhere.

I had also inadvertently added an extra request recording mechanism when
Graeme already implemented one in the graphql-mocks (which aren't really
mocks imho, but anyway). So used that and made it slightly simpler to
use.