feat: cache entity requests in engine-v2 (#1953)
Following on from #1931 - this adds caching on the entity requests to
subgraphs. These are a bit more complicated because one request will
usually map to more than one cache entry. We need to filter out any
representations that have a cache hit from the request to the subgraph,
and then make sure we use the cache entry for those during
deserialisation rather than trying to read them from the response.

I've made use of `serde_json::RawValue` in quite a few places to try to
make things more efficient - so we don't have to go through a bunch of
extra serde roundtrips.

Fixes GB-7159