chore: decouple redis pools from rate limiting (#1964)
I'm in the process of adding redis support to entity caching. Rate
limiting already uses redis, and ideally we want to be able to share the
connection pool between caching & rate limiting, assuming the user is
using the same redis settings for both.

This is a bit of prep work for that: extracting the generic redis
connection code into it's own module with its own structs, where they
were part of the rate limiting config/code before.

I've also added a `RedisPoolFactory` that is used to construct pools -
it'll return the same pool if you give it the same details twice. This
should make it easy to share connection pools without having to
explicity write code to compare a rate limiting config with an entity
caching config.

Part of GB-6894