feat: implement and hook up RedisEntityCache (#1973)
As per the title this finishes adding redis entity caching support.  

I haven't written any tests of this yet because the existing tests of
the gateway binary don't seem to call subgraphs, so there's a bit of
setup I'd need to do for that. Have added GB-7210 to handle that.

Fixes GB-6894