fix: include headers in entity caching keys
One of the things we want to avoid in entity caching is sharing cache entries
that shouldn't be shared: one way that could happen is if some kind of
authentication header was included in a subgraph request.  This handles that
case by adding subgraph headers into the entity caching key.

We probably still need to handle adding auth scopes and such into the caching
key, but that's a bigger job - this should handle a lot of the obvious cases.