feat(engine-v2): Add support for SSE subscriptions to subgraphs.
If the websocket URL is not defined we'll send a request with `Accept:
text/event-stream,application/json;q=0.9` to start a SSE subscriptions.
We may want to allow a different URL for this at some point, not sure.

I'm using the quick start of the JS `graphql-sse` library as a SSE
subgraph in the integration-tests.

I've changed the `Fetcher` trait to remove a bunch of allocations we're
doing for websockets, couldn't resist not removing those.