engine-v2: allow parameters in content-type header (#2023)
We accepted content-type: application/json;charset=utf-8 (and other
parameters after the semicolon) before the graphql-over-http
implementation. The spec says the content type should be
application/json, but not that parameters are forbidden. Us rejecting it
seems to be an error.

This commit makes the engine accept parameters in content type again.