gateway: remove subgraph urls from request errors (#2052)
Previously, when a subgraph was not reachable, the gateway would include
the url of the subgraph in the error. This behaviour is from one of our
dependencies, reqwest. We do not want to leak subgraph urls, so starting
with this commit, the url will not appear in the error anymore.

closes GB-7339