composition/engine-v2: fix introspection of defaults of enum type
The internal value type in engine-v2 has a variant for enum values that
contains an enum id. Since composition produced Value::Enum(id) where
id is a StringId, those were converted and rendered as strings.

There are two potential ways out of this situation: make engine-v2's value
type less precise, or make federated-graph's value type more precise. The
path taken in this PR is the second one. We introduce a Value::EnumValue()
variant that takes an EnumId, but retain the variant with a StringId inside
as Value::UnboundEnumValue. That is necessary because we don't have enum
values in the schema for enums that are used in directives. We don't do enough
analysis on directive definitions, but we still need to represent these values.

What we want to fix, and what is fixed in this PR, are enum values
as defaults in output field arguments and fields of input objects
(collectively: input value definitions). They now use the stricter Value::Enum
variant in compositionn and from_sdl().

One side effect of this change is that within a given enum, values cannot
be alphabetically sorted anymore in engine_v2::Schema, because that sorting
step would break the referencing ids in Values.

closes GB-7375