chore: split gateway binary tests into own crate
The gateway binary tests currently don't seem to run any actual subgraphs. I
want to add a test of redis entity caching to these tests, which is difficult
to do without access to some running subgraphs.  In our existing integration
tests we have quite a lot of support code that makes this really easy to do -
and I'd like to add something similar into the gateway binary tests.

However, doing this with the current setup felt a bit messy - there's no
obvious crate to put the stuff into, and adding submodules into a crates
integration tests felt a bit bad.

So: this PR splits the gateway-binary tests out into their own crate, where we
can put support code into the crate itself and then pull that into the tests.
Very much the same pattern as we use for the engine integration tests.  

I've also reworked some of the client code a little bit to be closer to the
patterns we use in the engine integration tests. It looked like the current
code was at least partially taken from either the CLI tests, which aren't
_wrong_ but I think the patterns we use in existing integration-tests are
better.