engine-v2: avoid re-parsing of federated SDL
This is a follow up to https://github.com/grafbase/grafbase/pull/2064.

We went into cycles of rendering and parsing federated SDL, creating the
conditions for the bug in render_federated_sdl() to manifest on engine start
up.

This change makes sure we parse only once, _either_ when constructing it
when instantiating Config directly, _or_ during deserialization.

closes GB-7388