federated-graph: start decluttering render_federated_sdl() (#2066)
While working on this morning's fix, I noticed this module could really
do with some love. Here's what this PR starts to address:

- Align render_federated_sdl() and render_api sdl() on `fmt::Formatter`
based APIs instead of writing directly to a `String`.
- Reduce the number of different ways we render directives. This PR
introduces a `DirectiveWriter` abstraction that should work in the whole
crate.
- Avoid the proliferation of types that only hold FederatedGraph, an
item and implement Display. Functions are more concise and easy to
follow.

These changes should not affect the output in any way. We have pretty
extensive tests in the graphql-federated-graph and graphql-composition
crates so I am fairly confident that this is the case.