chore: Improve docker build time
To read a full explanation: https://www.tweag.io/blog/2023-08-10-rust-static-link-with-mimalloc/

The short version is that the musl allocated behaves poorly in
multi-threaded scenarios. So the author patched musl to use mimalloc.
Now on my machine build time are comparable to those without docker
instead of spending an absurd amount of time in kernel threads.