Initial implementation of patching in graphql-schema-diff (#2072)
Until now, you could diff two GraphQL schemas with graphql-schema-diff,
and get the locations that changed, and what kind of change occurred. In
https://github.com/grafbase/grafbase/pull/2014, spans were added to the
changes to point out the exact content of the change, as well as a way
to resolve these spans to strings.

This PR implements patching with a `patch()` function that takes a diff,
the resolved spans for that diff, and a source schema, and applies the
diff to the schema. The format is intentionally designed to be easy to
store in a database table.

The implementation is not complete, there are missing pieces:

- directive definitions
- directives
- field arguments
- type extensions
- descriptions

I opted against finishing all of it now so we can make progress on the
features we are planning and fill in the gaps later. Also to get
feedback on patching earlier.