chore: remove current caching from CI
As far as I can tell this is not helping: it consumes far more space than
GitHub allocates us.  I have a few other things I'd like to try that might
be more effective, but it's hard to try them when this particular caching
constantly keeps us over the limit.