graphql-schema-diff: fix two bugs (#2101)
- Removing a root from a schema definition now works. We would
previously end up with declarations like:

 ```graphql
 schema {
   query: Query
   mutation:
 }
 ```

The reason being that graphql-schema-diff expresses removal as
`ChangeQueryType` (or mutation, subscription, respectively) with an
empty string as the type name. We may want to change that.

- In Paths::iter_exact(), we were taking the index of the path instead
of the index of the change, which was incorrect any time the changes
aren't sorted by path, which we can't guarantee.