chore: build before clippy (#2098)
Prior to this change the order of things was:

1. Debug clippy run of all pacakges
2. Tests
3. Debug build of binaries

Not certain but I suspect that this ends up duplicating some work - my
theory is that the cargo check that clippy runs is discarding some work
that the later cargo build needs to duplicate. It's also possible that
the test in the middle is picking up a slightly different set of
features to the original check and that is causing some rebuilds.

This PR changes the order to build, clippy, tests.

Anyway, as always it's hard to be certain but the two times I've ran
this in CI it ran in about 16 minutes - which is faster than the 20ish
we had before.