chore: more ci tweaks (#2106)
This PR fixes a couple of things I noticed:

1. Somehow the action I was using to download what-rust-changed was
consistently taking 3 minutes to run. It only took around a minute to
compile so 3 minutes to download was not an improvement. I've replaced
it with a custom action that just downloads it directly and runs in a
reasonable time.
2. I noticed that cargo was picking up the full list of targets from
`rust-toolchain.toml` and installing all of them in every CI job. Not a
big deal, but it was slowing things down by a little bit for every job.
Using `rustup override set` causes rust to ignore the toolchains file,
so I've done that.
3. Also extracted the rust install stuff into an action. (There is the
existing init-rust-job action I could put that in, but I want to leave
the existing stuff alone for now)