chore: add a workspace hack package & caching (#2093)
We make quite heavy use of feature flags in this repo, both with
external dependencies and with our own crates. These are great, but it's
very very easy to end up with slightly different feature sets if you're
compiling subsets of the repo. This can cause rebuilds of entire
dependency trees. I've certainly noticed this on my local machine when
running test suites for individual packages, and from my testing it also
happens on CI.

This PR adds a fix for that: a "workspace hack" package managed by
`cargo-hakari`. `cargo-hakari` looks at all the dependencies of all our
crates and adds them to the workspace hack package, with the full set of
features used everywhere. This ensures we always have a consistent set
of features enabled.

This has a few ramifications:

- It cuts down on unnecessary rebuilds locally.  A clear win
- It cuts down on unnecessary rebuilds in a CI job.  Also a clear win.
- It makes caching in CI easier, as the set of dependencies are always
unified.
- It does mean every CI run has to build all the dependencies - even if
they aren't directly used in the subset of packages we're building. Not
so good, but this can be mitigated by caching.

### Caching

With the workspace hack in place we need about 1gb of space to cache
each target. We have 4 targets so we can now fit a single builds cache
into GitHubs 10GB of cache space. However, if we change a few deps
concurrently we're easily going over 10GB. In my testing the GH cache is
also not consistently fast.

Buildjet & ubicloud both offer GHA compatible caches. When running on
ubicloud workers the GHA cache is a little slow, so
this PR enables the Buildjet cache for this. It was easier than
ubiclouds cache and they offer 20gb by default (and also appear open to
giving you more), whereas ubicloud only give 10 (and it's less clear
that they will give you more - though we could email them and ask).

I have tested builds on CI both with and without caching:

- For a full rebuild of the repo
- With this change but no caching I have seen 14 minute rebuilds (vs 16
or so without this change)
  - With caching enabled but nothing in the cache - 15m 20s
  - With caching enabled and a populated cache - 9m 31s
- For a build of _just_ gqlint (the smallest package in our repo in
terms of other packages it affects)
- With this change but no caching: 8 minutes (vs 4:40 without - quite
bad)
- I don't have full rebuild numbers for buildjet cache, but was 6m 27 to
build using GH caching and an empty cache
  - With a populated cache (on buildjet) - 4m 23

So provided we combine it with caching, this seems like a very clear
win.

I've also removed the quite ineffective old caching as part of this - it
was filling up our available cache storage basically every day, and
makes the caching that actually works harder. See #2099 for more
thinking on that.