chore(engine-v2): Add computed operation name & clean up graphql span
I'll need to add the operation name to the top-level HTTP span as we
want to filter on it in traces. Today it's computed in the telemetry-sink, which I can't do for
spans as there is no guarantee operation spans will be in the same batch
the HTTP one. At least I don't want to rely on it. So computing the name
here and using it for both metrics and traces propagated as
`grafbase.operation.computed_name`.

Instead of extensions method I'm using a simple wrapper for the GraphQL
span. Removed a few things for the subgraph ones, but adding them later
back differently. Also removed the tracing v1 tests, they're just too
painful to update every time, clickhouse integration-tests for traces
content are better adapted.